requirements.txt
streamlit
openai
python-dotenv
pdf2image